+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
--
+- \[Breaking\] Renamed `RedisHost` to `RedisConnectionString` in the config options for clarity
 
 ## [1.7.0] - 2020-03-30