+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
--
+- Update to .Net Core 3.1 (supported until Dec 2022) since 2.2 support ended last year
 
 ## [1.7.0] - 2020-03-30